응답 모델 분리
