+++ Boostcamp_iOS_BoxOffice/Controller/MovieTableViewController.swift
@@ -44,7 +44,6 @@ class MovieTableViewController: UIViewController {
         orderNumber = getOrderTypeUserDefaults()
         setNaviBarTitle(orderType: orderNumber)
         API.shared.requestMovies(orderType: orderNumber)
-        //requestMovies(orderType: orderNumber)
     }
     
     @objc func didReceiveMoviesNotification(_ noti: Notification) {
