+++ Boostcamp_iOS_BoxOffice/Model/Comments.swift
@@ -2,21 +2,11 @@
 //  Comments.swift
 //  Boostcamp_iOS_BoxOffice
 //
-//  Created by 이재은 on 14/12/2018.
-//  Copyright © 2018 이재은. All rights reserved.
+//  Created by Presto on 20/01/2019.
+//  Copyright © 2019 이재은. All rights reserved.
 //
 
 import Foundation
-
-struct CommentsAPIResponse: Codable {
-    let comments: [Comments]
-    let movieId: String
-    
-    enum CodingKeys: String, CodingKey {
-        case comments
-        case movieId = "movie_id"
-    }
-}
 
 struct Comments: Codable {
     let rating: Double
